+++ messytables/html.py
@@ -1,9 +1,18 @@
 from messytables.core import RowSet, TableSet, Cell, CoreProperties
 import lxml.html
 from collections import defaultdict
+import html5lib
+import xml.etree.ElementTree as etree
 
+def fromstring(s):
+    tb = html5lib.getTreeBuilder("lxml", implementation=etree)
+    p = html5lib.HTMLParser(tb, namespaceHTMLElements=False)
+    return p.parse(s)
 
 class HTMLTableSet(TableSet):
+    """
+    A TableSet from a HTML document.
+    """
     def __init__(self, fileobj=None, filename=None, window=None):
 
         if filename is not None:
@@ -14,7 +23,7 @@ class HTMLTableSet(TableSet):
             raise TypeError('You must provide one of filename or fileobj')
 
         self.htmltables = []
-        root = lxml.html.fromstring(fh.read())
+        root = fromstring(fh.read())
 
         # Grab tables that don't contain tables, remove from root, repeat.
         while True:
@@ -31,6 +40,9 @@ class HTMLTableSet(TableSet):
                 "other tables. This is a bug."  # avoid infinite loops
 
     def make_tables(self):
+        """
+        Return a listing of tables (as HTMLRowSets) in the table set.
+        """
         def rowset_name(rowset, table_index):
             return "Table {0} of {1}".format(table_index + 1,
                                              len(self.htmltables))
@@ -54,6 +66,9 @@ def insert_blank_cells(row, blanks):
 
 
 class HTMLRowSet(RowSet):
+    """
+    A RowSet representing a HTML table.
+    """
     def __init__(self, name, sheet, window=None):
         self.name = name
         self.sheet = sheet
@@ -139,7 +154,7 @@ class HTMLCell(Cell):
 
     def __init__(self, value=None, column=None, type=None, source=None):
         assert value is None
-        assert isinstance(source, lxml.html.HtmlElement)
+        assert isinstance(source, lxml.etree._Element)
         self._lxml = source
         if type is None:
             from messytables.types import StringType
@@ -166,6 +181,7 @@ class HTMLCell(Cell):
     def properties(self):
         return HTMLProperties(self._lxml)
 
+
 def text_from_element(elem):
     r"""
     >>> x = lxml.html.fromstring('''<td><center><span style="display:none; speak:none" class="sortkey">01879-07-01-0000</span>
@@ -184,10 +200,11 @@ def text_from_element(elem):
         else:
             cell_str = (x.text or '') + (x.tail or '')
         cell_str = cell_str.replace('\n', ' ').strip()
-        if x.tag =='br' or x.tag == 'p':
+        if x.tag == 'br' or x.tag == 'p':
             cell_str = '\n' + cell_str
         builder.append(cell_str)
     return ''.join(builder).strip()
+
 
 def is_invisible_text(elem):
     flag = False
@@ -198,11 +215,12 @@ def is_invisible_text(elem):
 
     return flag
 
+
 class HTMLProperties(CoreProperties):
     KEYS = ['_lxml', 'html', 'colspan', 'rowspan']
 
     def __init__(self, lxml_element):
-        if not isinstance(lxml_element, lxml.html.HtmlElement):
+        if not isinstance(lxml_element, lxml.etree._Element):
             raise TypeError("%r" % lxml_element)
         super(HTMLProperties, self).__init__()
         self.lxml_element = lxml_element