+++ test/test_read.py
@@ -3,21 +3,22 @@ import unittest
 
 from . import horror_fobj
 from nose.plugins.attrib import attr
-from nose.tools import (assert_equal, assert_greater_equal, assert_true,
-                        assert_raises)
+from nose.tools import assert_equal
+from nose.plugins.skip import SkipTest
+
 try:
-    # Python 2.6 doesn't provide assert_is_instance
-    from nose.tools import assert_is_instance
+    # Python 2.6 doesn't provide these functions
+    from nose.tools import assert_is_instance, assert_greater_equal
 except ImportError:
-    def assert_is_instance(obj, cls, msg=None):
-        assert_true(isinstance(obj, cls))
+    from shim26 import assert_is_instance, assert_greater_equal
 
 from messytables import (CSVTableSet, StringType, HTMLTableSet,
                          ZIPTableSet, XLSTableSet, XLSXTableSet, PDFTableSet,
                          ODSTableSet, headers_guess, headers_processor,
                          offset_processor, DateType, FloatType,
-                         IntegerType, rowset_as_jts,
-                         types_processor, type_guess, ReadError)
+                         IntegerType, BoolType, rowset_as_jts,
+                         types_processor, type_guess, ReadError,
+                         null_processor)
 import datetime
 
 
@@ -107,7 +108,7 @@ class ReadCsvTest(unittest.TestCase):
         table_set = CSVTableSet(fh)
         row_set = table_set.tables[0]
         types = type_guess(row_set.sample, strict=True)
-        expected_types = [IntegerType(), StringType(), IntegerType(),
+        expected_types = [IntegerType(), StringType(), BoolType(),
                           StringType()]
         assert_equal(types, expected_types)
 
@@ -123,6 +124,32 @@ class ReadCsvTest(unittest.TestCase):
 
         # we expect None for Integers and "" for empty strings in CSV
         assert [x.value for x in data[2]] == [3, "null", None, ""], data[2]
+
+    def test_null_process(self):
+        fh = horror_fobj('null.csv')
+        table_set = CSVTableSet(fh)
+        row_set = table_set.tables[0]
+        row_set.register_processor(null_processor(['null']))
+        data = list(row_set)
+
+        nones = [[x.value is None for x in row] for row in data]
+        assert_equal(nones[0], [False, True, False, False])
+        assert_equal(nones[1], [False, False, False, True])
+        assert_equal(nones[2], [False, True, False, False])
+
+        types = type_guess(row_set.sample, strict=True)
+        expected_types = [IntegerType(), BoolType(), BoolType(),
+                          BoolType()]
+        assert_equal(types, expected_types)
+
+        row_set.register_processor(types_processor(types))
+
+        # after applying the types, '' should become None for int columns
+        data = list(row_set)
+        nones = [[x.value is None for x in row] for row in data]
+        assert_equal(nones[0], [False, True, False, False])
+        assert_equal(nones[1], [False, False, False, True])
+        assert_equal(nones[2], [False, True, True, True])
 
     def test_read_encoded_csv(self):
         fh = horror_fobj('utf-16le_encoded.csv')
@@ -177,6 +204,18 @@ class ReadCsvTest(unittest.TestCase):
         assert 'foo' in data[12][0].column, data[12][0]
         assert 'Chirurgie' in data[12][0].value, data[12][0].value
 
+    def test_read_encoded_characters_csv(self):
+        fh = horror_fobj('characters.csv')
+        table_set = CSVTableSet(fh)
+        row_set = table_set.tables[0]
+        offset, headers = headers_guess(row_set.sample)
+        row_set.register_processor(headers_processor(headers))
+        row_set.register_processor(offset_processor(offset + 1))
+        data = list(row_set)
+        assert_equal(382, len(data))
+        assert_equal(data[0][2].value, u'雲嘉南濱海國家風景區管理處')
+        assert_equal(data[-1][2].value, u'沈光文紀念廳')
+
 
 class ReadZipTest(unittest.TestCase):
     def test_read_simple_zip(self):
@@ -220,6 +259,22 @@ class ReadSsvTest(unittest.TestCase):
 
         for row in list(row_set):
             assert_equal(3, len(row))
+            assert_equal(row[0].type, StringType())
+
+
+class ReadPsvTest(unittest.TestCase):
+    def test_read_simple_psv(self):
+        # pipe/vertical bar ("|") separated values
+        fh = horror_fobj('simple.psv')
+        table_set = CSVTableSet(fh)
+        row_set = table_set.tables[0]
+        assert_equal(29, len(list(row_set)))
+        row = list(row_set.sample)[0]
+        assert_equal(row[0].value, 'Year')
+        assert_equal(row[1].value, 'New dwellings')
+
+        for row in list(row_set):
+            assert_equal(6, len(row))
             assert_equal(row[0].type, StringType())
 
 
@@ -309,10 +364,13 @@ class ReadXlsTest(unittest.TestCase):
     # would be good to be able to.
     def test_attempt_read_encrypted_no_password_xls(self):
         fh = horror_fobj('encrypted_no_password.xls')
-        with assert_raises(ReadError) as cm:
-            XLSTableSet(fh)
         errmsg = "Can't read Excel file: XLRDError('Workbook is encrypted',)"
-        self.assertEqual(errmsg, str(cm.exception))
+        try:
+            XLSTableSet(fh)
+        except ReadError as e:
+            assert e.message == errmsg
+        else:
+            assert False, "Did not raise Read Error"
 
     def test_read_head_offset_excel(self):
         fh = horror_fobj('simple.xls')
@@ -492,7 +550,8 @@ class ReadPdfTest(unittest.TestCase):
                 PDFTableSet(fh)
             except ImportError:
                 # Optional library isn't installed. Skip the tests.
-                self.skipTest("pdftables is not installed, skipping PDF tests")
+                raise SkipTest("pdftables is not installed, skipping PDF tests")
+
 
     def test_read_simple_pdf(self):
         with horror_fobj('simple.pdf') as fh: